CSE-utilities/requirements.txt
gradio==4.44.1
sentence-transformers==3.0.1
faiss-cpu
scikit-learn==1.5.2
pypdf==5.0.1
docx2txt==0.8
openai==1.46.0
httpx==0.27.2